fix(scripts): split pilot helpers and drop pickle from tests for Codacy
Clear PR #57 Codacy gate (0 new issues max): split export_grok1_int8 and
route_preservation into sub-modules under Lizard file-NLOC limits, and replace
in-test pickle.dumps with committed golden fixtures so Bandit no longer flags
the unittest module. Offline generator remains for fixture regeneration.

Refs: #53 / RM-222
