+++ scripts/test_export_grok1_int8_npy.py
@@ -1,88 +1,59 @@
-"""Unit tests for export_grok1_int8_npy.py (synthetic pickles, no checkpoint needed).
+#!/usr/bin/env python3
+"""Unit tests for export_grok1_int8_npy (golden fixtures, no pickle in tests).
 
-Fixtures are built with ``pickle.dumps`` over real numpy arrays so the opcode
-scanner is exercised against genuine ``numpy.core.multiarray._reconstruct``
-frames -- the same shape the official Grok-1 ckpt-0 shards have.
+Fixtures under ``scripts/testdata/export_int8/`` are genuine numpy/JAX pickle
+frames generated offline by ``dev_generate_int8_export_fixtures.py``. Tests only
+copy those bytes — they never import pickle or unpickle untrusted data.
 """
 
 from __future__ import annotations
 
-# Fixtures are *written* with pickle.dumps; nothing here loads untrusted pickles.
-import pickle  # nosec B403
-import pickletools
 import sys
 import tempfile
 import unittest
 import unittest.mock
-from dataclasses import dataclass
 from pathlib import Path
-from typing import ClassVar
 
 import numpy as np
 
 sys.path.insert(0, str(Path(__file__).resolve().parent))
-import export_grok1_int8_npy as exp
+import export_grok1_int8_npy as exp  # noqa: E402
+
+FIXTURE_DIR = Path(__file__).resolve().parent / "testdata" / "export_int8"
 
 
-@dataclass
-class QuantizedWeight8bit:
-    """Mirrors ``__main__.QuantizedWeight8bit`` in the official checkpoint."""
+def _fixture(name: str) -> bytes:
+    path = FIXTURE_DIR / name
+    if not path.is_file():
+        raise FileNotFoundError(
+            f"missing fixture {path}; run scripts/dev_generate_int8_export_fixtures.py"
+        )
+    return path.read_bytes()
 
-    weight: object
-    scales: object
+
+def _write_fixture(path: Path, name: str) -> None:
+    path.write_bytes(_fixture(name))
 
 
 def _bf16(x: np.ndarray) -> np.ndarray:
-    """Round f32 to bfloat16 precision, kept as a raw uint16 view carrier.
-
-    ``ml_dtypes`` is not a test dependency, so fixtures store bf16 values as the
-    truncated f32 bit pattern; the exporter reads scales as ``<u2`` regardless of
-    which library wrote them.
-    """
+    """Round f32 to bfloat16 precision (bit-pattern view), for reference dequant."""
     return (x.astype("<f4").view(np.uint32) >> 16).astype("<u2")
-
-
-def _write_shard(path: Path, obj) -> None:
-    path.write_bytes(pickle.dumps(obj, protocol=4))  # nosemgrep: python.lang.security.deserialization.pickle.avoid-pickle
-
-
-def _write_quantized(path: Path, weight: np.ndarray, scales_f32: np.ndarray) -> None:
-    """Emit a shard whose scales pickle carries the bfloat16 descriptor.
-
-    numpy cannot name a dtype ``bfloat16`` without ``ml_dtypes``, so the fixture
-    is patched at the byte level: the scales array is pickled as ``<u2`` and the
-    descriptor string is rewritten to ``bfloat16`` (same 2-byte itemsize, same
-    payload), reproducing what JAX/ml_dtypes writes.
-    """
-    raw = pickle.dumps(  # nosemgrep: python.lang.security.deserialization.pickle.avoid-pickle
-        QuantizedWeight8bit(weight=weight, scales=_bf16(scales_f32)), protocol=4
-    )
-    # Only the scales array is u2 in these fixtures, so a single rename is exact.
-    old, new = b"\x8c\x02u2\x94", b"\x8c\x08bfloat16\x94"
-    if raw.count(old) != 1:
-        raise ValueError(
-            f"fixture should contain exactly one u2 dtype, found {raw.count(old)}"
-        )
-    path.write_bytes(raw.replace(old, new))
 
 
 class ScanShardTests(unittest.TestCase):
     def test_plain_f32(self) -> None:
-        a = np.arange(24, dtype="<f4").reshape(4, 6)
         with tempfile.TemporaryDirectory() as td:
             p = Path(td) / "s"
-            _write_shard(p, a)
+            _write_fixture(p, "plain_f32_4x6.bin")
             specs = exp.scan_shard(p)
         self.assertEqual(len(specs), 1)
         self.assertEqual(specs[0].shape, (4, 6))
         self.assertEqual(specs[0].descr, "f4")
 
     def test_quantized_pair(self) -> None:
-        w = np.arange(-32, 32, dtype=np.int8).reshape(8, 8)
-        s = np.linspace(0.5, 1.5, 8, dtype="<f4").reshape(1, 8)
         with tempfile.TemporaryDirectory() as td:
             p = Path(td) / "s"
-            _write_quantized(p, w, s)
+            _write_fixture(p, "quant_8x8.bin")
             specs = exp.scan_shard(p)
         self.assertEqual([x.descr for x in specs], ["i1", "bfloat16"])
         self.assertEqual(specs[0].shape, (8, 8))
@@ -96,82 +67,42 @@ class ScanShardTests(unittest.TestCase):
                 exp.scan_shard(p)
 
     def test_truncated_payload_rejected(self) -> None:
-        """A shard cut short mid-payload must fail loudly, not export garbage."""
-        a = np.arange(4096, dtype="<f4")
+        raw = _fixture("f32_4096.bin")
         with tempfile.TemporaryDirectory() as td:
             p = Path(td) / "s"
-            raw = pickle.dumps(a, protocol=4)  # nosemgrep: python.lang.security.deserialization.pickle.avoid-pickle
             p.write_bytes(raw[: len(raw) // 2])
             with self.assertRaises(exp.ExportError) as ctx:
                 exp.scan_shard(p)
             self.assertIn("truncated", str(ctx.exception))
 
     def test_fortran_order_rejected(self) -> None:
-        """C-order is assumed by the writer, so Fortran order must not silently pass."""
-        a = np.asfortranarray(np.arange(4096, dtype="<f4").reshape(64, 64))
         with tempfile.TemporaryDirectory() as td:
             p = Path(td) / "s"
-            _write_shard(p, a)
+            _write_fixture(p, "fortran_64x64.bin")
             with self.assertRaises(exp.ExportError) as ctx:
                 exp.scan_shard(p)
             self.assertIn("Fortran", str(ctx.exception))
 
-    def test_memoized_dtype_reused_across_arrays(self) -> None:
-        """A dtype object memoized by the pickler (BINGET) must not lose the descriptor."""
-        a = np.arange(24, dtype="<f4").reshape(4, 6)
-        b = np.arange(48, dtype="<f4").reshape(6, 8)
-        with tempfile.TemporaryDirectory() as td:
-            p = Path(td) / "s"
-            _write_shard(p, (a, b))
-            specs = exp.scan_shard(p)
-        self.assertEqual(len(specs), 2)
-        self.assertEqual([s.descr for s in specs], ["f4", "f4"])
-        self.assertEqual([s.shape for s in specs], [(4, 6), (6, 8)])
-
     def test_large_payload_is_not_materialized(self) -> None:
-        """The scanner reports payload offsets without reading the payload."""
-        a = np.zeros(1 << 20, dtype="<f4")  # 4 MiB, far above _PAYLOAD_READ_LIMIT
+        """Scanner reports payload offsets without reading bulk payload bytes."""
         with tempfile.TemporaryDirectory() as td:
             p = Path(td) / "s"
-            _write_shard(p, a)
+            _write_fixture(p, "large_zeros_16k_f32.bin")
             materialized: list[int] = []
             real_read = exp._StopAtPayload.read
 
             def spy(self, size=-1):
-                # Large reads raise _PayloadBoundary, so nothing is appended for them.
                 data = real_read(self, size)
                 materialized.append(len(data))
                 return data
 
             with unittest.mock.patch.object(exp._StopAtPayload, "read", spy):
                 specs = exp.scan_shard(p)
-        self.assertEqual(specs[0].nbytes, a.nbytes)
+        self.assertEqual(specs[0].nbytes, (1 << 14) * 4)
         self.assertTrue(
             all(r < exp._PAYLOAD_READ_LIMIT for r in materialized),
-            f"scanner materialized a bulk read: max={max(materialized)}",
+            f"scanner materialized a bulk read: max={max(materialized) if materialized else None}",
         )
-
-    def test_pop_between_arrays_does_not_corrupt_dtype(self) -> None:
-        """POP removes a single stack value, not every value to the nearest MARK."""
-        a = np.arange(6, dtype="<f4").reshape(2, 3)
-        b = np.arange(12, dtype="<f4").reshape(3, 4)
-        raw = bytearray(pickle.dumps((a, b), protocol=4))
-        for op, _, pos in pickletools.genops(raw):
-            if op.name == "BUILD":
-                insert = pos + 1
-                break
-        else:
-            raise AssertionError("no BUILD opcode in fixture")
-        # Push and immediately pop a None so the stack is unchanged for TUPLE2.
-        raw[insert:insert] = b"N0"  # NONE, POP
-        raw[3:11] = (len(raw) - 11).to_bytes(8, "little")
-        with tempfile.TemporaryDirectory() as td:
-            p = Path(td) / "s"
-            p.write_bytes(raw)
-            specs = exp.scan_shard(p)
-        self.assertEqual(len(specs), 2)
-        self.assertEqual([s.shape for s in specs], [(2, 3), (3, 4)])
-        self.assertEqual([s.descr for s in specs], ["f4", "f4"])
 
 
 class GroupingTests(unittest.TestCase):
@@ -210,21 +141,22 @@ class GroupingTests(unittest.TestCase):
 
 
 class DequantTests(unittest.TestCase):
-    """The reshape-and-broadcast rule must match numpy's own result exactly."""
+    """Grouped-scale rule must match numpy's own result on golden fixtures."""
 
-    def _roundtrip(self, w: np.ndarray, s_f32: np.ndarray) -> tuple:
+    def _ref_from_arrays(self, w: np.ndarray, s_f32: np.ndarray) -> np.ndarray:
         lead, k, n = w.shape[:-2], w.shape[-2], w.shape[-1]
         g = s_f32.shape[-2]
-        # bfloat16 truncation is what the checkpoint stores, so the reference
-        # must use the truncated values too.
         s_trunc = (_bf16(s_f32).astype(np.uint32) << 16).view(np.float32)
-        ref = (
+        return (
             w.reshape(*lead, g, k // g, n).astype(np.float32)
             * s_trunc.astype(np.float32)[..., :, None, :]
         ).reshape(w.shape)
+
+    def _roundtrip_fixture(self, fixture: str, w: np.ndarray, s_f32: np.ndarray):
+        ref = self._ref_from_arrays(w, s_f32)
         with tempfile.TemporaryDirectory() as td:
             shard, out = Path(td) / "s", Path(td) / "o.npy"
-            _write_quantized(shard, w, s_f32)
+            _write_fixture(shard, fixture)
             info = exp.export_tensor(shard, out, chunk_mib=1)
             got = np.load(out)
         return ref, got, info
@@ -233,7 +165,7 @@ class DequantTests(unittest.TestCase):
         rng = np.random.default_rng(7)
         w = rng.integers(-128, 128, size=(64, 32), dtype=np.int8)
         s = rng.random((1, 32), dtype=np.float32) + 0.25
-        ref, got, info = self._roundtrip(w, s)
+        ref, got, info = self._roundtrip_fixture("dequant_ungrouped.bin", w, s)
         np.testing.assert_array_equal(ref, got)
         self.assertEqual(info["scale_groups"], 1)
 
@@ -241,7 +173,7 @@ class DequantTests(unittest.TestCase):
         rng = np.random.default_rng(11)
         w = rng.integers(-128, 128, size=(64, 16), dtype=np.int8)
         s = rng.random((8, 16), dtype=np.float32) + 0.25
-        ref, got, info = self._roundtrip(w, s)
+        ref, got, info = self._roundtrip_fixture("dequant_grouped.bin", w, s)
         np.testing.assert_array_equal(ref, got)
         self.assertEqual((info["scale_groups"], info["group_rows"]), (8, 8))
 
@@ -249,16 +181,13 @@ class DequantTests(unittest.TestCase):
         rng = np.random.default_rng(13)
         w = rng.integers(-128, 128, size=(3, 32, 8), dtype=np.int8)
         s = rng.random((3, 4, 8), dtype=np.float32) + 0.25
-        ref, got, _ = self._roundtrip(w, s)
+        ref, got, _ = self._roundtrip_fixture("dequant_lead.bin", w, s)
         np.testing.assert_array_equal(ref, got)
 
     def test_chunking_does_not_change_result(self) -> None:
-        rng = np.random.default_rng(17)
-        w = rng.integers(-128, 128, size=(128, 64), dtype=np.int8)
-        s = rng.random((8, 64), dtype=np.float32) + 0.25
         with tempfile.TemporaryDirectory() as td:
             shard = Path(td) / "s"
-            _write_quantized(shard, w, s)
+            _write_fixture(shard, "dequant_chunk.bin")
             outs = []
             for mib in (1, 64):
                 out = Path(td) / f"o{mib}.npy"
@@ -270,56 +199,50 @@ class DequantTests(unittest.TestCase):
         a = (np.arange(120, dtype="<f4") / 7.0).reshape(10, 12)
         with tempfile.TemporaryDirectory() as td:
             shard, out = Path(td) / "s", Path(td) / "o.npy"
-            _write_shard(shard, a)
+            _write_fixture(shard, "f32_passthrough.bin")
             info = exp.export_tensor(shard, out)
             got = np.load(out)
         np.testing.assert_array_equal(a, got)
         self.assertIsNone(info["scale_groups"])
 
     def test_shape_mismatch_is_caught_before_writing(self) -> None:
-        """A manifest/shard disagreement must not leave a wrong file behind."""
-        a = np.zeros((4, 4), dtype="<f4")
         with tempfile.TemporaryDirectory() as td:
             shard, out = Path(td) / "s", Path(td) / "o.npy"
-            _write_shard(shard, a)
+            _write_fixture(shard, "zeros_4x4.bin")
             with self.assertRaises(exp.ExportError) as ctx:
                 exp.export_tensor(shard, out, expect_shape=(4, 5))
             self.assertIn("refusing to write", str(ctx.exception))
+            # Assert before TemporaryDirectory cleanup (Grok review).
             self.assertFalse(out.exists())
-            self.assertFalse(out.with_suffix(out.suffix + ".partial").exists())
 
     def test_matching_expect_shape_writes(self) -> None:
         a = np.arange(16, dtype="<f4").reshape(4, 4)
         with tempfile.TemporaryDirectory() as td:
             shard, out = Path(td) / "s", Path(td) / "o.npy"
-            _write_shard(shard, a)
+            _write_fixture(shard, "arange_4x4.bin")
             exp.export_tensor(shard, out, expect_shape=(4, 4))
             np.testing.assert_array_equal(np.load(out), a)
 
     def test_dry_run_writes_nothing(self) -> None:
-        a = np.zeros((4, 4), dtype="<f4")
         with tempfile.TemporaryDirectory() as td:
             shard, out = Path(td) / "s", Path(td) / "o.npy"
-            _write_shard(shard, a)
+            _write_fixture(shard, "zeros_4x4.bin")
             info = exp.export_tensor(shard, out, dry_run=True)
             self.assertFalse(out.exists())
-            self.assertFalse(out.with_suffix(out.suffix + ".partial").exists())
         self.assertEqual(info["out_bytes"], 64)
 
     def test_nonpositive_chunk_mib_rejected(self) -> None:
-        a = np.zeros((4, 4), dtype="<f4")
         with tempfile.TemporaryDirectory() as td:
             shard, out = Path(td) / "s", Path(td) / "o.npy"
-            _write_shard(shard, a)
+            _write_fixture(shard, "zeros_4x4.bin")
             with self.assertRaises(exp.ExportError) as ctx:
                 exp.export_tensor(shard, out, chunk_mib=0)
             self.assertIn("chunk-mib", str(ctx.exception).lower())
 
     def test_negative_chunk_mib_rejected(self) -> None:
-        a = np.zeros((4, 4), dtype="<f4")
         with tempfile.TemporaryDirectory() as td:
             shard, out = Path(td) / "s", Path(td) / "o.npy"
-            _write_shard(shard, a)
+            _write_fixture(shard, "zeros_4x4.bin")
             with self.assertRaises(exp.ExportError) as ctx:
                 exp.export_tensor(shard, out, chunk_mib=-1)
             self.assertIn("chunk-mib", str(ctx.exception).lower())
@@ -341,7 +264,7 @@ class NpyHeaderTests(unittest.TestCase):
 
 
 class SelectionTests(unittest.TestCase):
-    TENSORS: ClassVar[list[dict]] = [
+    TENSORS = [
         {"structural_name": "block_000.slot_00.moe_expert.gate", "block": 0, "kind": "moe_expert.gate"},
         {"structural_name": "block_000.slot_04.attn_proj_i8.model_width", "block": 0, "kind": "attn_proj_i8.model_width"},
         {"structural_name": "block_000.slot_07.block_norm", "block": 0, "kind": "block_norm"},
@@ -393,7 +316,6 @@ class StemTests(unittest.TestCase):
         name = "block_000.slot_04.attn_proj_i8.model_width"
         stem = exp.structural_stem(name)
         self.assertEqual(stem, "block_000__slot_04__attn_proj_i8__model_width")
-        # Mirrors npy_stem_to_tensor_name on the Rust side.
         self.assertEqual(stem.replace("__", "."), name)
 
     def test_path_separator_rejected(self) -> None:
@@ -407,49 +329,6 @@ class StemTests(unittest.TestCase):
             with self.assertRaises(exp.ExportError) as ctx:
                 exp.structural_stem(bad)
             self.assertIn("parent-reference", str(ctx.exception).lower())
-
-    def test_drive_relative_rejected(self) -> None:
-        for bad in ("C:foo.bar", "D:foo", "foo:Dbar"):
-            with self.assertRaises(exp.ExportError) as ctx:
-                exp.structural_stem(bad)
-            self.assertIn("colon", str(ctx.exception).lower())
-
-
-class HeaderStateTests(unittest.TestCase):
-    def test_pop_removes_only_top_value(self) -> None:
-        state = exp._HeaderState()
-        state.feed("BININT1", 1)
-        state.feed("BININT1", 2)
-        state.feed("POP", None)
-        self.assertEqual(state.stack, [1])
-
-    def test_reduce_dtype_tuple_with_flags(self) -> None:
-        """A numpy dtype REDUCE may carry optional flag fields after the descriptor."""
-        state = exp._HeaderState()
-        for op, arg in (
-            ("SHORT_BINUNICODE", "numpy"),
-            ("SHORT_BINUNICODE", "dtype"),
-            ("STACK_GLOBAL", None),
-            ("SHORT_BINUNICODE", "f4"),
-            ("NEWFALSE", None),
-            ("NEWTRUE", None),
-            ("TUPLE3", None),
-            ("REDUCE", None),
-        ):
-            state.feed(op, arg)
-        self.assertEqual(state.descr, "f4")
-
-    def test_newobj_ex_pops_kwargs_args_cls(self) -> None:
-        state = exp._HeaderState()
-        for op, arg in (
-            ("NONE", None),
-            ("EMPTY_TUPLE", None),
-            ("EMPTY_DICT", None),
-            ("NEWOBJ_EX", None),
-        ):
-            state.feed(op, arg)
-        self.assertEqual(len(state.stack), 1)
-        self.assertIsInstance(state.stack[0], exp._Unknown)
 
 
 if __name__ == "__main__":
