+++ scripts/testdata/export_int8/README.md
@@ -1,0 +1,4 @@
+# Golden pickle shards for export_grok1_int8 tests
+
+Regenerate with `python3 scripts/dev_generate_int8_export_fixtures.py`.
+Tests load these bytes only — no `pickle` import in the test module.
